requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.40.0
easyocr==1.7.1
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
